api/requirements.txt
Flask==3.0.3
GitPython
pandas
plotly.express
plotly.utils
us